+++ src/utils.rs
@@ -11,8 +11,8 @@ use windows::{
     core::*,
 };
 
-pub fn create_utf_string(string: &str) -> Vec<u16> {
-    let mut string_utf: Vec<u16> = string.encode_utf16().collect();
+pub fn create_utf_string(string: impl AsRef<str>) -> Vec<u16> {
+    let mut string_utf: Vec<u16> = string.as_ref().encode_utf16().collect();
     string_utf.push(0);
     string_utf
 }
@@ -72,10 +72,11 @@ pub fn kill(wanted_process_name: &str) {
         if Process32FirstW(snapshot, &mut entry).is_ok() {
             loop {
                 let process_name = String::from_utf16_lossy(&entry.szExeFile);
-                if process_name.contains(wanted_process_name) && entry.th32ProcessID != current_pid {
-                    if let Ok(process) = OpenProcess(PROCESS_TERMINATE, false, entry.th32ProcessID) {
-                        TerminateProcess(process, 0).ok();
-                    }
+                if process_name.contains(wanted_process_name)
+                    && entry.th32ProcessID != current_pid
+                    && let Ok(process) = OpenProcess(PROCESS_TERMINATE, false, entry.th32ProcessID)
+                {
+                    TerminateProcess(process, 0).ok();
                 }
                 if Process32NextW(snapshot, &mut entry).is_err() {
                     break;
@@ -90,7 +91,7 @@ pub fn set_cpu_throttling(webview: &ICoreWebView2, value: f32) {
         webview
             .CallDevToolsProtocolMethod(
                 w!("Emulation.setCPUThrottlingRate"),
-                PCWSTR(create_utf_string(&format!("{{\"rate\":{}}}", value)).as_ptr()),
+                PCWSTR(create_utf_string(format!("{{\"rate\":{}}}", value)).as_ptr()),
                 None,
             )
             .ok();
